spotify-agent/requirement.txt
# Core dependencies
google-adk>=0.1.0
spotipy>=2.23.0
python-dotenv>=1.0.0

# Utilities
requests>=2.31.0